Add interface for ReadTenantService
